Refactor Launch component, remove ConnectedView
The move to using modern hooks/redux allows us to removed the
`ConnectedView` component, removing a bit of complexity. In fact, the
redux docs suggest removing
`connect`/`mapStateToProps`/`mapDispatchToProps`.

We also take this time to refactor out our last class component in
`Launch` and move that over to a function component`